drm: bridge: adv7511: Add flag to disable dynamically switch lane count
Not all platforms can support the use of 3 lanes due to how they calculate
path, and for higher resolutions, 2-lanes may not be enough.  To address
this, create an optional flag called "adi,fixed-lanes" to tell the
adv7533/7535 that the connected platform cannot dynamically switch the
lane count.  When enabled, the number of lane will always be equal to
adi,dsi-lanes regardless of the resolution or pixel clock.

Signed-off-by: Adam Ford <aford173@gmail.com>