arm64: imx8mn: Configure power domains for LCDIF and DSI
The i.MX8M Nano uses an LCDIF to generate an image which
is then routed thought a DSI controller.  Both of these
items are gated by power domain controllers. Set the clock
parents in the power domains before the clocks are enabled,
since they cannot be re-parented once they are running.

Signed-off-by: Adam Ford <aford173@gmail.com>