arm64: dts: imx8mp-beacon-kit:  Move sai3 to Audio_PLL1
The Beacon board has an LVDS display that cannot get a proper clock
rate from Video_PLL if the DSI is operational due to the way the
clock divides, so the LVDS needs to use an alternative clock.
The Audio_PLL2 clock can feed the LDB.

Because the clock rated neede for the LDB isn't divisable by
the clock rate needed by SAI3, move SAI3 to use Audio_PLL1,
and reconfigure the CODEC to use a 12MHz fixed clock.

Because these clocks are no longer in sync with each other,
reconfigure the SAI3 to be the master since the
CODEC can internally compensate when fed a fixed clock
reference, even if it is not an even multiple of the desired reate.

This now leave AUDIO_PLL2 completely free for the LDB without
compromising the audio sound from the codec.

Signed-off-by: Adam Ford <aford173@gmail.com>