arm64: dts: imx8mp-beacon: Enable LVDS-1
Beacon has an display that can connect to one of the LVDS ports on
the baseboard.  The display requires a 30MHz clock to display
properly, and the LDB needs to run at 7x that. With the audio CODEC
now moved to the AUDIO_PLL1, the AUDIO_PLL2 is now available to
source the LDB at 210MHz and the DISP_PIX2.

Signed-off-by: Adam Ford <aford173@gmail.com>