net: ethernet: davinci_emac: Support optional reset driver
On the AM3517, there is a register that must write a 0 then a 1
to pull the davinci_emac out of reset.  Currently there is a
function in mach-omap2/pdata-quirks.c to work around this.

Now that there is a proper reset driver for it, update the
davinci_emac driver to look for the device tree reference
and do this work instead of assumming pdata-quirks did it.

Signed-off-by: Adam Ford <aford173@gmail.com>