reset: reset-ti-am35xx:  Add support for new driver
The AM35xx has a single register with different bits that
must be set to 1 then cleared to 0 before some peripherals
can be used.  Currently there is a function in pdata-quirks.c
that manually does this.  Instead of this older method, create
a reset driver which can be called from those various peripherals
that need it and can be assigned via the device tree.

Signed-off-by: Adam Ford <aford173@gmail.com>