usb: musb: am35xx: Enable AM35xx MUSB Glue
The AM3517 has a variant of the Mentor Graphics USB OTG controller
that is different from both the DA850 and the AM33xx despite their
names being similar.

The Interrupts are handled via a specialized syscon register, and
the USB PHY uses a unique PHY driver specific to the AM35xx.

Signed-off-by: Adam Ford <aford173@gmail.com>