dmaengine: ti: cppi41: Add prelimiary support for AM35xx CPPA41
The CPPI41 DMA controller is similar to the DA850 and the AM33xx,
but it has a different offset than either of the those.  It uses
a few registers similar to the DA850, but it has more queues, yet
not quite like the AM33xx.  Update the CPPA41 drivers to support
these unique features.

Signed-off-by: Adam Ford <aford173@gmail.com>