phy: freescale: fsl-samsung-hdmi: Stop searching when exact match is found
There are a series of for-loops which check various values of P and S
for the integer divder PLL.  The for loops search all entries and use
the one closest to the nominal, but it searches through all for loops
even if the nominal is achieved.  Ending when the nominal value is
foudn stops wasting time since it'll not find find a better value.

Signed-off-by: Adam Ford <aford173@gmail.com>