phy: freescale: fsl-samsung-hdmi: Reorder integer divider before fractional
In preparation to properly calculate the fractional divider, it's necessary
to calculate the integer clock first.  This patch just changes the order
from checking the look-up table first.  Once the order is switched,
the phy_clk_round_rate will use the closest value to the desired
between the fractional and the integer.

Signed-off-by: Adam Ford <aford173@gmail.com>