phy: freescale: fsl-samsung-hdmi: Enable Fractional divider
When the integer divder is close but not exact, there is a fractional
divider that uses a series of variables to offset the integer divider
to can get the desired clock closer to the desired value.

This then allows for the removal of the look-up table and use the
dynamically calculated values.

Signed-off-by: Adam Ford <aford173@gmail.com>