drm/bridge: sn65dsi83: Rework ti-sn65dsi83 to play with NXP DSIM
There is a weird catch-22 with the NXP DSIM driver and the n65dsi83
where one driver returns with EPROBE_DEFER because the other driver
is not loaded.  The other driver then does the same.  This repeats
until the system gives up.  Rework the driver so it registers without
attaching to the DSI host by delaying the attach until later.