phy: freescale: fsl-samsung-hdmi: Add support for fractional divider
There is some fairly time-consuming calcuations to determine the
best values for a series of variables used for setting the
fractional divider for the phy-fsl-samsung-hdmi.  This can
be helpful if there are unusual clock requirements that
are not already available in the lookup table.

The driver work flow is now:

1.  Search the lookup table for fractional divider entries
    and use it if it's a match.
2.  Calculate the integer divider entries and use it
    if it's a match.
3.  (new)  Calculate the fractional divider and use it
    in combination with the integer values if it's a match.

Signed-off-by: Adam Ford <aford173@gmail.com>